move createCombinedSystemMessage method to composable
